fix(*): renamed local common module to avoid shadowing agr_file_generator common module
